Move files and change namespaces to make code more readable
