+++ src/Order.Service.Tests/OrderServiceTests.cs
@@ -5,6 +5,7 @@ using Moq;
 using NUnit.Framework;
 using Order.Data;
 using Order.Data.Entities;
+using Order.Data.Repositories;
 using Order.Data.Specifications.Evaluators;
 using Order.Model;
 using Order.Service.Exceptions;
