+++ src/Order.Service/OrderService.cs
@@ -1,5 +1,9 @@
 ﻿using Order.Data;
+using Order.Data.Repositories;
 using Order.Data.Specifications;
+using Order.Data.Specifications.OrderSummarySpec;
+using Order.Data.Specifications.ProductSpec;
+using Order.Data.Specifications.StatusSpec;
 using Order.Model;
 using Order.Service.Exceptions;
 using Order.Service.Specifications;
